12b: Split input.go and subagent.go to restore 1000-line cap
Local smoke-test of the ci-check pipeline flagged two files that drifted
past 1000 lines during the Phase 4/7 option-struct refactors. input.go
(1013) splits into input.go + input_keys.go along the existing
key-dispatch boundary; subagent.go (1112) splits into subagent.go +
subagent_output.go along result-formatting / output-persistence. All
three rules are now green locally.
